Add config to disable the `to_s` override in Ruby core classes
This will allow users to take advantage early of Ruby 3.1 optimization
in string interpolation.
