+++ activesupport/lib/active_support/core_ext/date.rb
@@ -4,4 +4,5 @@ require "active_support/core_ext/date/acts_like"
 require "active_support/core_ext/date/blank"
 require "active_support/core_ext/date/calculations"
 require "active_support/core_ext/date/conversions"
+require "active_support/core_ext/date/deprecated_conversions" unless ENV["RAILS_DISABLE_DEPRECATED_TO_S_CONVERSION"]
 require "active_support/core_ext/date/zones"
