+++ activesupport/lib/active_support/core_ext/date_time.rb
@@ -5,3 +5,4 @@ require "active_support/core_ext/date_time/blank"
 require "active_support/core_ext/date_time/calculations"
 require "active_support/core_ext/date_time/compatibility"
 require "active_support/core_ext/date_time/conversions"
+require "active_support/core_ext/date_time/deprecated_conversions" unless ENV["RAILS_DISABLE_DEPRECATED_TO_S_CONVERSION"]
