+++ activesupport/lib/active_support/core_ext/range.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "active_support/core_ext/range/conversions"
+require "active_support/core_ext/range/deprecated_conversions" unless ENV["RAILS_DISABLE_DEPRECATED_TO_S_CONVERSION"]
 require "active_support/core_ext/range/compare_range"
 require "active_support/core_ext/range/overlaps"
 require "active_support/core_ext/range/each"
